requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
geographiclib==2.1
geopy==2.4.1
idna==3.10
mysqlclient==2.2.7
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.2
pillow==11.3.0
PyJWT==2.10.1
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
